requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dependency-injector==4.41.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql==0.0.3
mysqlclient==2.2.4
orjson==3.10.3
packaging==24.0
pluggy==1.5.0
prompt_toolkit==3.0.45
pybase62==1.0.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
random-strings==0.1.1
redis==5.0.4
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uvicorn==0.30.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
